Move msgfame to it's own module
